pyproject.toml
[project]
name = "azureml-parallel-validation"
version = "0.1.0"
description = "Reference sample: custom Docker validation framework as an AzureML parallel job"
requires-python = ">=3.10,<3.13"
license = "MIT"

dependencies = [
    "mltable>=1.6.1",
    "azureml-fsspec>=1.3.1",
    "pandas>=2.2.2,<3",
    "setuptools<81",  # required by mltable (imports pkg_resources, removed in 81+)
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pyyaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
